Add Black-Scholes option pricing implementation in test.cpp
- Implemented a command-line interface for calculating call and put option prices using the Black-Scholes model.
- Added input prompts for current price, strike price, expiration date, risk-free interest rate, and volatility.
- Integrated call and put pricing functions from the Black_Scholes class.

+++ Black_Scholes.h
@@ -4,21 +4,21 @@
 class Black_Scholes
 {
 public:
-    double current_price;           // S: current stock/asset price
-    double strike_price;            // K: strike price
-    double time_to_expiration;      // T: time in years
-    double risk_free_interest_rate; // r: annual risk-free rate as decimal
-    double volatility;              // σ: annual volatility as decimal
+    float current_price;           // S: current stock/asset price
+    float strike_price;            // K: strike price
+    float time_to_expiration;      // T: time in years
+    float risk_free_interest_rate; // r: annual risk-free rate as decimal
+    float volatility;              // σ: annual volatility as decimal
 
-    double call(double, double, double, double, double);
+    float call(float, float, float, float, float);
 
-    double put(double, double, double, double, double);
+    float put(float, float, float, float, float);
 
-    double d1(double, double, double, double, double, double);
+    float d1(float, float, float, float, float, float);
 
-    double d2(double, double, double, double, double);
+    float d2(float, float, float, float, float);
 
-    double N(double x);
+    float N(float x);
 };
 
 #endif
